Change the behaviour how the amount of total clicks is handled.
The amount of total clicks is now handled here in the server side.
Whenever a client clicks the button, the client no longer sends the
total amount of clicks from his/her state to other clients. Instead, when
the click event happens, the total amount of clicks is increased here in
the server side.

Whenever a client disconnects from the app, the socket listener checks if that
client was the last client using the app. If it was, the total amount of clicks
is sent to the database. Whenever a client connects to the app, if the new client
is the only client in the app, the amount of clicks is fetched from the database.